[refactor] move device detection to platform module
