+++ lmcache/__init__.py
@@ -8,7 +8,12 @@ import types
 
 # First Party
 from lmcache.logging import init_logger
-from lmcache.v1.platform.device_ext import DeviceExt
+
+# --------------------------
+# Device detection
+# --------------------------
+from lmcache.v1.platform.device_detection import torch_dev as torch_dev
+from lmcache.v1.platform.device_detection import torch_device_type as torch_device_type
 
 try:
     # First Party
@@ -17,61 +22,8 @@ except ImportError:
     __version__ = "unknown"
 
 logger = init_logger(__name__)
-# Standard
 
 __all__ = ["__version__", "torch_dev", "torch_device_type"]
-
-
-# --------------------------
-# Device detection
-# --------------------------
-def _detect_device() -> tuple[Any, str]:
-    """
-    Detect the available accelerator and return the corresponding torch
-    device module and device type string.
-
-    Returns:
-        tuple[Any, str]: A tuple of (torch_device_module, device_type_string),
-            e.g. ``(torch.cuda, "cuda")``, ``(torch.musa, "musa")``, or
-            ``(torch.xpu, "xpu")``.
-    """
-    try:
-        # Third Party
-        import torch
-    except ImportError:
-        return None, "cpu"  # fallback，CLI-only
-
-    if hasattr(torch, "musa") and torch.musa.is_available():  # type: ignore[attr-defined]
-        logger.info("MUSA device is available. Using MUSA for LMCache engine.")
-        return torch.musa, "musa"  # type: ignore[attr-defined]
-    elif hasattr(torch, "xpu") and torch.xpu.is_available():
-        return torch.xpu, "xpu"
-    elif hasattr(torch, "hpu") and torch.hpu.is_available():
-        return torch.hpu, "hpu"
-    elif torch.cuda.is_available():
-        return torch.cuda, "cuda"
-    else:
-        # First Party
-        from lmcache.v1.platform.cpu.stub_cpu_device import StubCPUDevice
-
-        # Fallback: always return torch, cpu as stub
-        return StubCPUDevice("cpu"), "cpu"
-
-
-torch_dev, torch_device_type = _detect_device()
-
-logger.info(" torch_dev=%s, torch_device_type=%s", torch_dev, torch_device_type)
-
-
-# Attach the DeviceExt instance as ``torch_dev.ext``.  This monkey-patches a
-# standard torch module (e.g. ``torch.cuda``) with a custom attribute that does
-# not exist in the original module.  The ``# type: ignore[attr-defined]`` suppresses
-# the expected mypy/pyright "attr-defined" error from this intentional extension.
-if torch_dev is not None:
-    torch_dev.ext = DeviceExt(torch_device_type)  # type: ignore[attr-defined]
-else:
-    logger.warning("torch_dev is None, skipping DeviceExt initialization.")
-    pass
 
 
 # --------------------------
