+++ lmcache/v1/platform/device_detection.py
@@ -1,0 +1,69 @@
+# SPDX-License-Identifier: Apache-2.0
+"""Device detection logic for LMCache.
+
+This module detects the available hardware accelerator and exposes
+:data:`torch_dev` and :data:`torch_device_type` as module-level
+singletons.  It is intentionally a leaf module -- it does NOT import
+from ``lmcache.__init__`` or ``lmcache.v1.platform.__init__`` to
+avoid circular dependencies.
+
+Other modules should import from here (or from the re-export in
+``lmcache.__init__``) rather than performing their own detection.
+"""
+
+# Standard
+from typing import Any
+
+# First Party
+from lmcache.logging import init_logger
+
+logger = init_logger(__name__)
+
+
+def _detect_device() -> tuple[Any, str]:
+    """Detect the available accelerator and return the torch device module.
+
+    Returns:
+        tuple[Any, str]: A tuple of (torch_device_module, device_type_string),
+            e.g. ``(torch.cuda, "cuda")``, ``(torch.musa, "musa")``, or
+            ``(torch.xpu, "xpu")``.  When torch is not installed (CLI-only
+            mode), returns ``(None, "cpu")``.
+    """
+    try:
+        # Third Party
+        import torch
+    except ImportError:
+        return None, "cpu"  # fallback for CLI-only environments
+
+    if hasattr(torch, "musa") and torch.musa.is_available():  # type: ignore[attr-defined]
+        logger.info("MUSA device is available. Using MUSA for LMCache engine.")
+        return torch.musa, "musa"  # type: ignore[attr-defined]
+    elif hasattr(torch, "xpu") and torch.xpu.is_available():
+        return torch.xpu, "xpu"
+    elif hasattr(torch, "hpu") and torch.hpu.is_available():
+        return torch.hpu, "hpu"
+    elif torch.cuda.is_available():
+        return torch.cuda, "cuda"
+    else:
+        # First Party
+        from lmcache.v1.platform.cpu.stub_cpu_device import StubCPUDevice
+
+        # Fallback: return a stub that mimics a torch device module
+        return StubCPUDevice("cpu"), "cpu"
+
+
+torch_dev, torch_device_type = _detect_device()
+
+logger.info("torch_dev=%s, torch_device_type=%s", torch_dev, torch_device_type)
+
+# Attach the DeviceExt instance as ``torch_dev.ext``.  This monkey-patches a
+# standard torch module (e.g. ``torch.cuda``) with a custom attribute that does
+# not exist in the original module.  The ``# type: ignore[attr-defined]``
+# suppresses the expected mypy/pyright "attr-defined" error.
+if torch_dev is not None:
+    # First Party
+    from lmcache.v1.platform.device_ext import DeviceExt
+
+    torch_dev.ext = DeviceExt(torch_device_type)  # type: ignore[attr-defined]
+else:
+    logger.warning("torch_dev is None, skipping DeviceExt initialization.")
